Refactored the network layer a bit and added some state machines
The network layer and client model are now a bit cleaner from the
consumer side but at the expense of the caller never seeing the ACK
flows.  The ACK flows are now hidden behind state machines and only
unsolicited (inbound publish and pingresp) are ever delivered to the
client (on the server this will include disconnect, connect and
subscribe).  A little code cleanup but it's still just thrown together
haphazard.